backend/requirements/prod.txt
# Django & DRF packages
django==4.2.16
djangorestframework==3.16.0
drf-spectacular==0.28.0
djangorestframework-simplejwt==5.5.0

# Others
python-dotenv==1.1.0
psycopg2-binary==2.9.10
